server-deploy/server/requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Django Channels for WebSocket
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# MySQL
mysqlclient>=2.2.0
PyMySQL>=1.1.0

# WeChat Login
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Production server
gunicorn>=21.2.0
